feat: config 객체 삭제 및 시작 포인트에 설정 메소드 구현
